+++ tests/tensor_ops_test.py
@@ -558,40 +558,6 @@ def test_detach(with_detach):
 
 
 @helpers.printCapfdOnExit
-def test_inplace_inputs(capfd):
-    class Model(torch.nn.Module):
-        def forward(self, x):
-            if isinstance(x, (tuple, list)):
-                x[0].add_(1)
-            elif isinstance(x, (dict)):
-                x['input'] += 1
-            else:
-                x += 1
-
-            return x
-
-    poptorch_model = poptorch.inferenceModel(Model())
-    tensor_in = torch.Tensor([1.0])
-    assert poptorch_model(tensor_in) == 2.0
-    assert tensor_in == 1.0
-
-    list_in = (torch.Tensor([1.0]), )
-    assert poptorch_model(tensor_in)[0] == 2.0
-    assert list_in[0] == 1.0
-
-    log = helpers.LogChecker(capfd)
-    msg = ("An input tensor is modified in-place by the model. This is " +
-           "not supported on the IPU and the input will remain unchanged. " +
-           "This applies to all in-place operations such as \"+=\", \"*=\" " +
-           "or those ending in \"_\". To avoid this warning, please use the " +
-           "non in-place alternatives such as \"x = x + 1\" instead of " +
-           "\"x += 1\" or the operation not ending in \"_\" matching the " +
-           "in-place variant, on all model inputs.")
-
-    log.assert_contains(msg)
-
-
-@helpers.printCapfdOnExit
 def test_requires_grad_true(capfd):
     model = torch.nn.Linear(1, 1)
     poptorch_model = poptorch.inferenceModel(model)