+++ docs/user_guide/changelog.rst
@@ -21,6 +21,7 @@ New features
 - Deprecated ``poptorch.Options.Training.accumulationReductionType`` in favour of ``poptorch.Options.Training.accumulationAndReplicationReductionType``
 - Add support for ``torch.scatter_add``
 - Add ``poptorch.Options.Precision.enableFloatingPointExceptions`` to control floating point exception behavior
+- Add support for inplace changes to inputs.
 - Add option to log the number of IPU cycles used in executing the main graph
 - Deprecated ``runningVarianceAlwaysFloat`` in favour of ``runningStatisticsAlwaysFloat`` in ``poptorch.Options.Precision``,
   as this new option computes both the running mean and variance in FP32 when this option is set to `True`.